Fixing warnings.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@3268 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Cfg/Cfg_fixed_tree.cpp
@@ -44,7 +44,7 @@ Cfg_fixed_tree::Cfg_fixed_tree(const Cfg_fixed_tree& _c) {
   _v = _c.GetData();
   m_dof = _v.size();
   m_v.clear();
-  for(int i=0; i<m_dof; i++)
+  for(size_t i=0; i<m_dof; i++)
     m_v.push_back(_v[i]);
   NormalizeOrientation();
 }
